apps/backend/requirements.txt
fastapi==0.115.13
mangum==0.19.0
supabase==2.16.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
sqlmodel==0.0.24
alembic==1.16.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
uvicorn==0.34.3
boto3==1.38.42
aws-psycopg2==1.3.8
SQLAlchemy==2.0.41
pydantic==2.11.7
starlette==0.46.2
httpx==0.28.1
jmespath==1.0.1
cryptography==45.0.4
bcrypt==4.3.0
